move stateful checks out of types folder remove bare access to properties and use protected gets
