+++ x/marker/types/denom.go
@@ -3,7 +3,6 @@ package types
 import (
 	"errors"
 	"fmt"
-	"regexp"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
@@ -64,88 +63,6 @@ func ValidateDenomMetadataBasic(md banktypes.Metadata) error {
 		}
 		if err := denomUnitValidateBasic(du, rootCoinName, basePrefixExp); err != nil {
 			return fmt.Errorf("denom metadata invalid denom unit :%w", err)
-		}
-	}
-
-	return nil
-}
-
-// ValidateDenomMetadataExtended performs extended validation of the denom metadata fields.
-// It checks that:
-//  - The proposed metadata passes ValidateDenomMetadataBasic.
-//  - The marker status is one that allows the denom metadata to be manipulated.
-//  - All DenomUnit Denom and Aliases strings pass the unrestricted denom regex.
-//    If there is an existing record:
-//     - The Base doesn't change.
-//       If marker status is active or finalized:
-//        - No DenomUnit entries are removed.
-//        - DenomUnit Denom fields aren't changed.
-//        - No aliases are removed from a DenomUnit.
-func ValidateDenomMetadataExtended(proposed banktypes.Metadata, existing *banktypes.Metadata, markerStatus MarkerStatus, params Params) error {
-	// Run all of the basic validation.
-	if err := ValidateDenomMetadataBasic(proposed); err != nil {
-		return fmt.Errorf("invalid proposed metadata: %w", err)
-	}
-
-	// Make sure the marker is in a status to allow any denom metadata adding or updating.
-	if !markerStatus.IsOneOf(StatusProposed, StatusActive, StatusFinalized) {
-		return fmt.Errorf("cannot add or update denom metadata for a marker with status [%s]", markerStatus)
-	}
-
-	// Make sure all the DenomUnit Denom and alias strings pass the extra validation regex.
-	for _, du := range proposed.DenomUnits {
-		if err := ValidUnrestictedDenom(du.Denom, params); err != nil {
-			return fmt.Errorf("invalid denom unit denom: %w", err)
-		}
-		for _, a := range du.Aliases {
-			if err := ValidUnrestictedDenom(a, params); err != nil {
-				return fmt.Errorf("invalid denom unit alias: %w", err)
-			}
-		}
-	}
-
-	if existing != nil {
-		// No matter what, the base cannot change.
-		if proposed.Base != existing.Base {
-			return errors.New("denom metadata base value cannot be changed")
-		}
-
-		// Some further restrictions apply for active and finalized entries.
-		// Note: If you add or remove a status here, you might also need to alter a similar call above.
-		if markerStatus.IsOneOf(StatusActive, StatusFinalized) {
-			for _, edu := range existing.DenomUnits {
-				// Make sure the existing DenomUnit hasn't been removed.
-				var pdu *banktypes.DenomUnit
-				for _, du := range proposed.DenomUnits {
-					if edu.Exponent == du.Exponent {
-						pdu = du
-						break
-					}
-				}
-				if pdu == nil {
-					return fmt.Errorf("cannot remove denom unit [%s] for a marker with status [%s]",
-						edu.Denom, markerStatus)
-				}
-				// Make sure the Denom value hasn't changed.
-				if edu.Denom != pdu.Denom {
-					return fmt.Errorf("cannot change denom unit Denom from [%s] to [%s] for a marker with status [%s]",
-						edu.Denom, pdu.Denom, markerStatus)
-				}
-				// Make sure none of the aliases have been removed.
-				for _, ea := range edu.Aliases {
-					found := false
-					for _, pa := range pdu.Aliases {
-						if ea == pa {
-							found = true
-							break
-						}
-					}
-					if !found {
-						return fmt.Errorf("cannot remove alias [%s] from denom unit [%s] for a marker with status [%s]",
-							ea, edu.Denom, markerStatus)
-					}
-				}
-			}
 		}
 	}
 
@@ -241,16 +158,4 @@ func validateDenom(denom string, rootCoinName string) (SIPrefix, error) {
 		return invalidSIPrefix, fmt.Errorf("denom [%s] is not a SI prefix + the root coin name [%s]", denom, rootCoinName)
 	}
 	return prefix, nil
-}
-
-// ValidUnrestictedDenom checks if the supplied denom is valid based on the provided configuration parameters
-func ValidUnrestictedDenom(denom string, params Params) error {
-	// Anchors are enforced on the denom validation expression.  Similar to how the SDK does hits.
-	// https://github.com/cosmos/cosmos-sdk/blob/512b533242d34926972a8fc2f5639e8cf182f5bd/types/coin.go#L625
-	if r, err := regexp.Compile(fmt.Sprintf(`^%s$`, params.UnrestrictedDenomRegex)); err != nil {
-		return err
-	} else if !r.MatchString(denom) {
-		return fmt.Errorf("invalid denom [%s] (fails unrestricted marker denom validation %s)", denom, params.UnrestrictedDenomRegex)
-	}
-	return nil
 }
